// src/Sound/ISoundOptions.d.ts
import { IFadeOptions } from '../Fade/IFadeOptions';
import { INodeOptions } from '../Node/INodeOptions';
import { ITaggableNodeOptions } from '../Node/ITaggableNodeOptions';
export interface ISoundOptions extends INodeOptions, ITaggableNodeOptions {
    audioElement?: HTMLAudioElement;
    buffer?: AudioBuffer;
    fade?: boolean | IFadeOptions;
    loop?: boolean;
    trackPosition?: number;
    getManagerVolume?(): number;
}